add name accessor
