+++ packages/utilities/fast-benchmarks/src/runFile.js
@@ -1,11 +1,6 @@
 /* eslint-disable no-undef */
 import { getTestStartName, measureMemory, updateComplete } from "../utils/index.js";
 (async () => {
-    // const container = document.getElementById("container");
-    // const create = () => {
-    //     const el = document.createElement("x-app");
-    //     return container.appendChild(el);
-    // };
     const run = async () => {
         const start = getTestStartName(test);
         performance.mark(start);
@@ -14,7 +9,6 @@ import { getTestStartName, measureMemory, updateComplete } from "../utils/index.
         }
         await updateComplete();
         performance.measure(test, start);
-        // destroy(container);
     };
     await run();
     measureMemory();
